+++ core/src/main/kotlin/no/elg/hex/screens/LevelCreationScreen.kt
@@ -35,12 +35,15 @@ import ktx.async.KtxAsync
 import ktx.async.MainDispatcher
 import ktx.scene2d.actor
 import ktx.scene2d.horizontalGroup
+import ktx.scene2d.tooltip
 import ktx.scene2d.vis.spinner
 import ktx.scene2d.vis.visCheckBox
 import ktx.scene2d.vis.visImage
 import ktx.scene2d.vis.visLabel
 import ktx.scene2d.vis.visTable
 import ktx.scene2d.vis.visTextButton
+import ktx.scene2d.vis.visTextTooltip
+import ktx.scene2d.vis.visTooltip
 import no.elg.hex.Hex
 import no.elg.hex.island.Island
 import no.elg.hex.island.IslandGeneration
@@ -322,22 +325,22 @@ class LevelCreationScreen :
         space(10f)
         visTextButton("Create island") {
           disableables.add(this)
-          onClick {
-            if (this.isDisabled) return@onClick
+          onInteract(stage, allPressedKeyShortcuts = intArrayOf(Keys.ENTER)) {
+            if (this.isDisabled) return@onInteract
             createNewIsland(false)
           }
         }
 
         visTextButton("Instantly create") {
           disableables.add(this)
-          onClick {
-            if (this.isDisabled) return@onClick
+          onInteract(stage, allPressedKeyShortcuts = intArrayOf(Keys.SPACE)) {
+            if (this.isDisabled) return@onInteract
             createNewIsland(true)
           }
         }
 
         visTextButton("Cancel") {
-          onInteract(stage, keyShortcut = intArrayOf(Keys.ESCAPE)) {
+          onInteract(stage, allPressedKeyShortcuts = intArrayOf(Keys.ESCAPE)) {
             Hex.screen = LevelSelectScreen()
           }
         }